--- app/src/main/java/com/example/toss_and/presentation/TempHomeFragment.kt
@@ -1,14 +1,0 @@
-package com.example.toss_and.presentation
-
-import android.os.Bundle
-import android.view.View
-import com.example.toss_and.R
-import com.example.toss_and.databinding.FragmentTemphomeBinding
-import com.example.toss_and.util.base.BindingFragment
-
-class TempHomeFragment : BindingFragment<FragmentTemphomeBinding>(R.layout.fragment_temphome) {
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
