+++ app/src/main/java/com/example/toss_and/presentation/main/all/AllFragment.kt
@@ -1,12 +1,12 @@
-package com.example.toss_and.presentation
+package com.example.toss_and.presentation.main.all
 
 import android.os.Bundle
 import android.view.View
 import com.example.toss_and.R
-import com.example.toss_and.databinding.FragmentStockBinding
+import com.example.toss_and.databinding.FragmentAllBinding
 import com.example.toss_and.util.base.BindingFragment
 
-class StockFragment : BindingFragment<FragmentStockBinding>(R.layout.fragment_stock) {
+class AllFragment : BindingFragment<FragmentAllBinding>(R.layout.fragment_all) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
