+++ app/src/main/java/com/example/toss_and/presentation/main/benefit/BenefitFragment.kt
@@ -1,12 +1,12 @@
-package com.example.toss_and.presentation
+package com.example.toss_and.presentation.main.benefit
 
 import android.os.Bundle
 import android.view.View
 import com.example.toss_and.R
-import com.example.toss_and.databinding.FragmentAllBinding
+import com.example.toss_and.databinding.FragmentBenefitBinding
 import com.example.toss_and.util.base.BindingFragment
 
-class AllFragment : BindingFragment<FragmentAllBinding>(R.layout.fragment_all) {
+class BenefitFragment : BindingFragment<FragmentBenefitBinding>(R.layout.fragment_benefit) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
