+++ app/src/main/java/com/example/toss_and/presentation/main/stock/StockFragment.kt
@@ -1,12 +1,12 @@
-package com.example.toss_and.presentation
+package com.example.toss_and.presentation.main.stock
 
 import android.os.Bundle
 import android.view.View
 import com.example.toss_and.R
-import com.example.toss_and.databinding.FragmentBenefitBinding
+import com.example.toss_and.databinding.FragmentStockBinding
 import com.example.toss_and.util.base.BindingFragment
 
-class BenefitFragment : BindingFragment<FragmentBenefitBinding>(R.layout.fragment_benefit) {
+class StockFragment : BindingFragment<FragmentStockBinding>(R.layout.fragment_stock) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
