+++ app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".presentation.HomeFragment">
+    tools:context=".presentation.main.home.HomeFragment">
 
     <data>
